requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cssutils==2.2.0
dnspython==2.1.0
filelock==3.0.12
Flask==1.1.2
idna==2.10
importlib-metadata==3.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==23.0.1
lxml==4.6.3
MarkupSafe==1.1.1
premailer==3.7.0
py3-validate-email==1.0.0
pywin32-ctypes==0.2.0
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
Werkzeug==1.0.1
yagmail==0.14.247
zipp==3.4.1
gunicorn
